#346 Rename `combine_compile.py` and `combine_pp.py`
